requirements.txt
absl-py==0.10.0
astunparse==1.6.3
attrs==20.3.0
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cycler==0.10.0
Cython==0.29.21
dataclasses==0.8
dill==0.3.3
dm-tree==0.1.5
future==0.18.2
gast==0.3.3
gin-config==0.4.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-cloud-bigquery==2.5.0
google-cloud-core==1.4.4
google-crc32c==1.0.0
google-pasta==0.2.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
h5py==2.10.0
httplib2==0.18.1
idna==2.10
importlib-metadata==3.1.1
importlib-resources==3.3.0
joblib==1.0.0
kaggle==1.5.10
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
numpy==1.18.5
oauthlib==3.1.0
opencv-python-headless==4.4.0.46
opt-einsum==3.3.0
pandas==1.1.4
Pillow==8.0.1
promise==2.3
proto-plus==1.11.0
protobuf==3.14.0
psutil==5.7.3
py-cpuinfo==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.4
sentencepiece==0.1.94
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-addons==0.11.2
tensorflow-datasets==4.1.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.10.0
tensorflow-metadata==0.25.0
tensorflow-model-optimization==0.5.0
termcolor==1.1.0
text-unidecode==1.3
tf-models-official==2.3.0
tf-slim==1.1.0
threadpoolctl==2.1.0
tqdm==4.54.0
typeguard==2.10.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0